requirements.txt
# ── Core framework ──────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
streamlit==1.41.1
requests==2.32.3
urllib3==1.26.20
charset_normalizer==3.4.0

# ── Database ────────────────────────────────────────────────────────────
sqlalchemy==2.0.36

# ── Authentication ──────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
bcrypt==4.2.1

# ── AI / LLM (Groq - free tier) ────────────────────────────────────────
langchain==0.3.14
langchain-groq>=0.3,<1
langchain-community==0.3.14
langchain-huggingface>=0.1,<1
sentence-transformers>=3.0

# ── Vector store ────────────────────────────────────────────────────────
faiss-cpu==1.9.0.post1

# ── PDF processing ──────────────────────────────────────────────────────
pypdf==5.1.0

# ── Export (PDF + DOCX) ────────────────────────────────────────────────
reportlab==4.2.5
python-docx==1.1.2
